Start es server before search specs only
* Prevents need for env var when running single specs
