+++ spec/support/request_spec_helper.rb
@@ -2,23 +2,21 @@ require 'rack/test'
 require 'rspec'
 require "savon/mock/spec_helper"
 
-require File.dirname(__FILE__) + "/../lib/app.rb"
+require File.dirname(__FILE__) + "/../../lib/app.rb"
 
 ENV['RACK_ENV'] = 'test'
 
-module RequestMixin
+module RequestSpecHelper
   include Rack::Test::Methods
   def app
     App
   end
 
   def send_json(method, url, data)
-    send(method, url, data.to_json, { "CONTENT_TYPE" => "application/json" })
+    send(method, url, data.to_json, { "Content-Type" => "application/json" })
   end
 end
 
 RSpec.configure do |config|
-  config.include RequestMixin
+  config.include RequestSpecHelper
 end
-
-
